Fixing case where only one interface is reported
Netconf responses that have only one element return a hash, not an
array containing one hash. I've added a check to get_interfaces to
handle this case correctly, but we'll probably need to add checks for
this behavior accross the entire library. The other option is to turn
on ForceArray in xml::simple, but this will require changing how we
currently access data in responses.